chore: Moves EmptyState component to components package
