+++ packages/components/src/organisms/EmptyState/index.ts
@@ -1,0 +1,2 @@
+export { default } from './EmptyState'
+export type { EmptyStateProps } from './EmptyState'
