--- packages/core/src/components/ui/EmptyState/index.ts
@@ -1,1 +1,0 @@
-export { default } from './EmptyState'
